Move DeprecatedVaultFilterService to vault folder
